refactor(app-router): fix layout, types
