+++ packages/next-js-app/app/loading.tsx
@@ -1,0 +1,9 @@
+'use client'
+
+import { Skeleton } from '@mui/material'
+
+const ProjectLoading = () => {
+  return <Skeleton></Skeleton>
+}
+
+export default ProjectLoading
